Swatting warnings

+++ splus.tech/WebS.h
@@ -20,7 +20,7 @@ public:
 protected:
 	virtual void onClientConnected(int clientSocket);
 	virtual void onClientDisconnected(int clientSocket);
-	virtual void onMessageReceived(int clientSocket, const char* msg, int length) override;
+	virtual void onMessageReceived(int clientSocket, const char* msg, size_t length) override;
 };
 
 #endif
